refactor(rust): extract shared freshell-ws integration-test harness into tests/common
